json and javascript
was informed by fahrel to added the functionality of the login webpage, i made a json file to format and save pre existing usernames/passwords. also made the JavaScript that fetches the json file and adds responses and improves overall security of the webpage. also creates objects that can be saved and makes sure the credentials are valid